requirements.txt
tensorflow==2.15.0
tensorflow_hub
Pillow
numpy
matplotlib
